step4 refactoring
