+++ module-api/src/main/java/kitchenpos/Application.java
@@ -2,9 +2,7 @@ package kitchenpos;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-@EnableJpaAuditing
 @SpringBootApplication
 public class Application {
     public static void main(String[] args) {
