+++ module-api/src/main/java/kitchenpos/menu/ui/MenuGroupRestController.java
@@ -26,14 +26,12 @@ public class MenuGroupRestController {
         final MenuGroupResponse created = menuGroupService.create(menuGroupRequest.toMenuGroup());
         final URI uri = URI.create("/api/menu-groups/" + created.getId());
         return ResponseEntity.created(uri)
-                .body(created)
-                ;
+                .body(created);
     }
 
     @GetMapping("/api/menu-groups")
     public ResponseEntity<List<MenuGroupResponse>> list() {
         return ResponseEntity.ok()
-                .body(menuGroupService.list())
-                ;
+                .body(menuGroupService.list());
     }
 }
