+++ module-api/src/main/java/kitchenpos/menu/ui/MenuRestController.java
@@ -1,5 +1,6 @@
 package kitchenpos.menu.ui;
 
+import kitchenpos.menu.Menu;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.dto.MenuRequest;
 import kitchenpos.menu.dto.MenuResponse;
@@ -25,14 +26,12 @@ public class MenuRestController {
         final MenuResponse created = menuService.create(menuRequest.toMenu());
         final URI uri = URI.create("/api/menus/" + created.getId());
         return ResponseEntity.created(uri)
-                .body(created)
-                ;
+                .body(created);
     }
 
     @GetMapping("/api/menus")
     public ResponseEntity<List<Menu>> list() {
         return ResponseEntity.ok()
-                .body(menuService.list())
-                ;
+                .body(menuService.list());
     }
 }
