+++ module-api/src/main/java/kitchenpos/order/applicatioin/TableGroupService.java
@@ -1,5 +1,6 @@
 package kitchenpos.order.applicatioin;
 
+import kitchenpos.order.*;
 import kitchenpos.order.domain.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
