+++ module-api/src/test/java/kitchenpos/product/ui/application/ProductServiceTest.java
@@ -1,5 +1,6 @@
-package kitchenpos.application;
+package kitchenpos.product.ui.application;
 
+import kitchenpos.product.application.ProductService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
