+++ module-domain/build.gradle
@@ -1,8 +1,5 @@
-bootJar { enabled = false }
-jar { enabled = true }
+apply plugin: 'java-library'
 
 dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    runtimeOnly 'com.h2database:h2'
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    api('org.springframework.boot:spring-boot-starter-data-jpa')
 }
