+++ module-domain/src/main/java/kitchenpos/product/ProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package kitchenpos.product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
