+++ module-domain/src/test/java/kitchenpos/ModuleDomainApplicationTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ApplicationTest {
+class ModuleDomainApplicationTest {
     @Test
     void contextLoads() {
     }
