+++ module-domain/src/test/java/kitchenpos/product/ProductRepositoryTest.java
@@ -1,7 +1,5 @@
 package kitchenpos.product;
 
-import kitchenpos.Product;
-import kitchenpos.ProductRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
